Separate export for OpenLayers integration
