+++ src/ts/flatgeobuf.ts
@@ -1,5 +1,4 @@
 export * as generic from './generic.js';
-export * as ol from './ol.js';
 export * as geojson from './geojson.js';
 
 export { Column } from './flat-geobuf/column.js';
